+++ src/envelope/DecayEnvelope.h
@@ -5,22 +5,22 @@ class DecayEnvelope {
 
 private:
 	int    t;
-	double a;
-	double b;
-	double releaseTime;
-	double curvature;
+	float a;
+	float b;
+	float releaseTime;
+	float curvature;
 	bool   stopped;
 
 	void update() {
 		// FIXME: if parameters changes while envelope playing,
 		//        there will be an audio glich.
 		//        process should only use previous output value, not t.
-		a = (2 - 4 * curvature) / (double)(releaseTime * releaseTime);
-		b = (4 * curvature - 3) / (double)releaseTime;
+		a = (2 - 4 * curvature) / (float)(releaseTime * releaseTime);
+		b = (4 * curvature - 3) / (float)releaseTime;
 	};
 
 public:
-	double out;
+	float out;
 
 	DecayEnvelope() {
 		releaseTime = 500; // TODO: release time in ms, not in tick
@@ -35,17 +35,17 @@ public:
 		t = 0;
 	};
 
-    double tic() {
+    float tic() {
 		if (stopped) return 0;
 		if (t++ > releaseTime) {
 			stopped = true;
 			out = 0;
 		}
-		out = a * (double)(t * t) + b * (double)t + 1;
+		out = a * (float)(t * t) + b * (float)t + 1;
 		return out;
 	};
 
-	double tic(double trig) {
+	float tic(float trig) {
 		if (trig > 0.8) {
 			stopped = false;
 			t = 0;
@@ -55,11 +55,11 @@ public:
 			stopped = true;
 			out = 0;
 		}
-		out = a * (double)(t * t) + b * (double)t + 1;
+		out = a * (float)(t * t) + b * (float)t + 1;
 		return out;
 	};
 
-	void setCurvature(double c) {
+	void setCurvature(float c) {
 		curvature = c;
 		update();
 	};
