+++ src/oscillator/OscTri.h
@@ -8,13 +8,13 @@ class OscTri : public Oscillator {
 private:
 
 public:
-	double width;
+	float width;
 
 	OscTri() {
 		width = 0.5;
 	};
 
-    double tic() {
+    float tic() {
 		pos += freq / SAMPLE_RATE;
 		if (pos > 1) pos -= 1;
 		out = 1 - 2 * ((pos < width) ? pos / width : 1 - (pos - width) / (1 - width));
