+++ src/utils/audioUtils.cpp
@@ -2,6 +2,6 @@
 #include <math.h>
 
 
-double noteToFreq(double noteNumber) {
+float noteToFreq(float noteNumber) {
 	return 440 * pow(2, (noteNumber - 69) / 12);
 }
